moved common logic to ManipulationTask
